refactor: extract compile helper shared by build-binary and build-all
